application/requirements.txt
flask
requests
gevent
pillow
Keras
tensorflow
